Update database schema
The database name was changed to match the one used on RDS.

Foreign key constraints must be initially deferred so that multiple tables can be updated in a transaction.

The location_id column of the events table may be NULL, to reflect that calendar events don't require a
location.

Many columns were given uniqueness constraints, to ensure rows are not duplicated. In the contacts
table, this requires entries to not be nullable, because the SQL standard specifies that NULL values are
unique, not equal.

Forgotten tables were added to the table grants for the scraper and alexaskill roles. Additional grants for those roles to use the sequences created by the primary keys were also added.