Merge the initial scraper for populating the DB
Database credentials are stored in a separate, ignored file.

Putting the secrets in a separate file makes it easier to work on
code that isn't sensitive while avoiding committing sensitive info.

A sample of the secrets file was created as a template for the
secrets file to be used in production.

Amazon's public root trust certificate is included as a convenience.