Allow schema to be specified in secrets.py
This moves the schema name out of the SQL queries and into a
variable in secrets.py.  The search path is set to this schema when
the cursor is created.

Making the schema easily configurable means we can easily use one
database for multiple people to do testing, by each person working
in a different schema.