Introduce a simple tool for managing test schema
Database/tools/test-schema-tool.py automates the tasks of configuring
secrets.py and listing, creating, and dropping schema for testing.

The tool creates a schema based on Database/schema.sql but with a
different name.

The only dependency is psycopg2.

READMEs have been added/updated to reflect the existence and utility
of this new tool.