Fix scraper, ditch calendar UIDs
After adding the UNIQUE constraint to the summary and start columns
in the events table, duplicate events were getting rejected as
intended.  However, this caused the event_id for those rows to also
go away, which caused the event_categories relationships for the
missing rows to violate foreign key constraints.

The solution is to look up the event_id from the database in addition
to the category_id, instead of using the event_id of the event being
processed.  This causes a duplicate row to be generated for the
table, which gets gracefully ignored, instead of aborting the whole
transaction.

Since we have deduplicated the events appearing on multiple calendars,
the UIDs from the calendars are not needed.