Flush progress indicators in scraper.py
This fixes the problem where nothing would appear to be happening
for a while, and then all the dots appeared at once when the script
finished.  Now they appear gradually, as each event is processed.

Also adds an affirmation at the end of the output and makes progress
visible only when run as a script.