Script the creation of scraper deployment package
The package.sh script must be run on an EC2 Amazon Linux x86_64
instance.  It compiles postgresql static libraries and compiles
psycopg2 using this static lib, and with support for SSL.  This is
necessary because the AWS Lambda execution environment does not have
PostgreSQL installed and does not provide a way for specifying
dependencies, so they must be bundled in the deployment package.

There are a few libraries in pip that appear to be static versions
of psycopg2, but unfortunately the do not seem to work as expected.

The method for compiling psycopg2 with static libpg was inspired by
https://github.com/jkehler/awslambda-psycopg2