Refactor CalendarConversation and CalendarHelper
This change breaks the long if-else chain for dispatching to the
appropriate handlers down into smaller, separate switch statements.  The
division helps reinforce what state the session is expected to be for
each handler.

Magic strings have been replaced with appropriately named constants in
several places.  The state integer is now an enum with descriptive
identifiers.

Many repeated response formats have been reduced into a few simple helper
functions.

One handler has been converted to the new PreparedStatement style of
querying the database, as a demonstration of the correct usage.

The CalendarHelper class has been promoted to be responsible for all
special SSML formatting logic for calendar fields.

A generic event formatter has been added which takes an array of the
desired calendar fields, a Map of events, and an index, returning a
String of the fields described with appropriate surrounding grammar and
SSML markup.

The listEvents() method now takes an array of fields as well, internally
relying on the new formatter to build the final string.

`<speak>` tags are no longer added by any tags in this class, leaving
the responsibility on the caller to appropriately build their final
response SSML.