Avoid NullPointerException when no session state
In new sessions, there is no stored state, so the ATTRIB_STATEID session
attribute does not exist.  `session.getAttribute(ATTRIB_STATEID)`
returns `null` in this case.  Rather than calling `.toString()` on the
potentially `null` state object and then parsing that `String` to get a
`SessionState` enum value, we can just cast the `Object` directly to a
`SessionState` and handle the `null` case in the following `switch`
block (it falls into the `default` case).