Add parsing for Last-Modified header
This parser is dependent on the specific formatting of the header.
If the format returned by the server changes, or if different servers
using different formats are added to the list of calendars, support
for these additional formats will have to be added.  Currently, a
failure to parse the header will fail safe and default to parsing
the calendar rather than skipping it.