Update last_updated when scraping a calendar
This is the final piece to enable the scraper to only update calendars
when they have been modified, if a 'Last-Modified' header is given by
the server.

Some refactoring was required in order to have access to both the url_id
and the database cursor in the same place.  `get_records()` now iterates
over a single given calendar, and the looping over URLs happens int
`populate_database()`.

As a minor enhancement, the scraper now prints a '+' symbol for each
calendar URL as it is encountered.  This allows the user to see that
progress is occurring when no dots are being printed for the individual
events (because the calendar has been skipped).